pyproject.toml
[tool.poetry]
name = "deployment-scripts"
version = "0.1.0"
description = ""
authors = ["Prabal Pathak <prabal01pathak@gmail.com>"]
readme = "README.md"
packages = [{include = "deployment_scripts"}]

[tool.poetry.dependencies]
python = "^3.11"
azure-mgmt-resource = "^23.0.1"
azure-mgmt-web = "^7.1.0"
azure-identity = "^1.13.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
